# study-service/src/main/resources/openapi/study-api.openapi.yml
openapi: 3.0.0
info:
  title: Study API
  version: 0.0.1
servers:
  - url: http://localhost:8083/api/study
paths:
...
  /group/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Group Information
      tags:
        - Group
      operationId: getGroupById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyGroupOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Group By Id
      tags:
        - Group
      operationId: deleteGroupById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /group:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Groups Information
      tags:
        - Group
      operationId: getAllGroups
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyGroupOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Group
      tags:
        - Group
      operationId: createGroup
      requestBody:
        description: Group Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyGroupOpenAPI'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /group/user:
    post:
      security:
        - bearerAuth: [ ]
      summary: Add user in group
      tags:
        - Group
      operationId: addUserInGroup
      requestBody:
        description: User id, group id
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInGroupRequestOpenApi'
      responses:
        '201':
          description: Added
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete user from group
      tags:
        - Group
      operationId: deleteUserFromGroup
      requestBody:
        description: User id, group id
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInGroupRequestOpenApi'
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /group/users:
    post:
      security:
        - bearerAuth: [ ]
      summary: Add users in group
      tags:
        - Group
      operationId: addUsersInGroup
      requestBody:
        description: User's ids, group's ids
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/UserInGroupRequestOpenApi'
      responses:
        '201':
          description: Added
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete users in group
      tags:
        - Group
      operationId: deleteUsersFromGroup
      requestBody:
        description: User's ids, group's ids
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/UserInGroupRequestOpenApi'
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
...
  /attendance-statistics/group/{group_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Group's Attendance Statistics Information
      tags:
        - AttendanceStatistics
      operationId: getStatisticsByGroupId
      parameters:
        - in: path
          name: group_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceStatisticsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /materials:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Materials Information
      tags:
        - Materials
      operationId: getAllMaterials
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Material
      tags:
        - Materials
      operationId: createMaterials
      requestBody:
        description: Materials Object
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StudyMaterialsBody'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
    put:
      security:
        - bearerAuth: [ ]
      summary: Update a Materials
      tags:
        - Materials
      operationId: updateMaterials
      requestBody:
        description: Materials Object
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StudyMaterialsBody'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /materials/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Materials Information
      tags:
        - Materials
      operationId: getMaterialsById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Material by Id
      tags:
        - Materials
      operationId: deleteMaterials
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /homework/materials/{materialsId}:
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Homework
      tags:
        - Homework
      operationId: createHomework
      parameters:
        - in:
            path
          name: materialsId
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: Homework Object
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SaveHomeworkRequestOpenApi'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
...
  /subject/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subject Information
      tags:
        - Subject
      operationId: getSubjectById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Subject by Id
      tags:
        - Subject
      operationId: deleteSubject
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /subject/group/{group_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Subjects Information By Group Id
      tags:
        - Subject
      operationId: getSubjectsByGroupId
      parameters:
        - in: path
          name: group_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubjectOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /materials/subject/{subject_id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get All Materials by Subject Id
      tags:
        - Materials
      operationId: getMaterialsBySubjectId
      parameters:
        - in: path
          name: subject_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /download/materials/{fileId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Download File by It's Id
      tags:
        - File Download
      operationId: downloadMaterialsFileById
      parameters:
        - in: path
          name: fileId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StudyMaterialsBody:
      type: object
      required:
        - studyMaterials
      properties:
        studyMaterials:
          $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        files:
          type: array
          items:
            type: string
            format: binary
            description: Files with Materials
...
    StudyGroupOpenAPI:
      type: object
      required:
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        studentsIds:
          type: array
          description: studentsIDs
          items:
            type: integer
            format: int64
    SubjectOpenAPI:
      type: object
      required:
        - name
        - studyGroupsIds
        - creatorId
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        studyGroupsIds:
          type: array
          items:
            type: integer
            format: int64
        studyMaterialsIds:
          type: array
          items:
            type: integer
            format: int64
        scheduleIds:
          type: array
          items:
            type: integer
            format: int64
        creatorId:
          type: integer
          format: int64
        editorsIds:
          type: array
          items:
            type: integer
            format: int64
...
    StudyMaterialsType:
      type: string
      enum:
        - MATERIAL
        - TASK
    StudyMaterialsOpenAPI:
      type: object
      required:
        - type
        - subjectId
        - materialsText
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/StudyMaterialsType"
        materialsText:
          type: string
        subjectId:
          type: integer
          format: int64
        homeworkIds:
          type: array
          items:
            type: integer
            format: int64
        filesId:
          type: array
          items:
            type: integer
            format: int64
        filesNames:
          type: array
          items:
            type: string
    HomeworkOpenAPI:
      type: object
      required:
        - taskId
        - studentId
        - date
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        grade:
          type: integer
        comment:
          type: string
        filesNames:
          type: array
          items:
            type: string
        fileIds:
          type: array
          items:
            type: integer
            format: int64
    SaveHomeworkRequestOpenApi:
      type: object
      required:
        - homework
        - file
      properties:
        homework:
          $ref: "#/components/schemas/HomeworkOpenAPI"
        file:
          type: array
          items:
            type: string
            format: binary
            description: File with homework
    UserInGroupRequestOpenApi:
      type: object
      required:
        - userId
        - groupId
      properties:
        userId:
          type: integer
          format: int64
        groupId:
          type: integer
          format: int64
...
    AttendanceStatisticsOpenApi:
      type: object
      required:
        - scheduleId
        - userId
        - status
      properties:
        scheduleId:
           type: integer
           format: int64
        status:
          $ref: "#/components/schemas/AttendanceStatusOpenApi"
        studentId:
          type: integer
          format: int64
    AttendanceStatusOpenApi:
      type: string
      enum:
        - ATTENDED
        - MISSED
...